Allow the SEAPATH_LVM profile to be loaded by default
Without this patch, the deployment needs a user to be physically
connecter to the server (with keyboard and screen) to choose the
profile.
With this commit, the SEAPATH_LVM profile is loaded by default.
An upcoming commit will allow the user to preset an ip address on a
specific NIC so that network is available right after deployement.

Signed-off-by: Florent CARLI <florent.carli@rte-france.com>