Set defaut networking
This commit allow the user to configure ip addressing to a specified NIC
so that remote access to the deployed server is possible right after
deployment.
Along with the "no menu" feature that uses the SEAPATH_LVM profile by
default, this allows a host to be deployed without even having to plug a
screen and keyboard.

Signed-off-by: Florent CARLI <florent.carli@rte-france.com>