Reboot automatically after deployment
A precedent commit allowed the deployment to start without the user to
even to press enter.
Another one allowed the user to preconfigure a network configuration.
With all that it should be possible to deploy a node without even having
to plug a screen and keyboard on the machine.
However there still is the need to press ENTER at the end of the
deployment, just before the reboot.
This commit fixes this.

The change is in the grub configuration, which was included in the
docker image. This means to be taken into account, this fix would
require the user to rebuild the docker image, which is painful.
In order to change this, this commit also removes the /etc/fai configuration from
the docker image and mount it as a container volume. That way any change
in the /etc/fai configuration is effective without having to rebuild the
image.

Signed-off-by: Florent CARLI <florent.carli@rte-france.com>