rename SEAPATH class to SEAPATH_COMMON
build_debian_iso is intended to build the seapath iso but also a default
vm for SEAPATH in the future.
This VM will have many similarities with the current iso, but with less
packages and less scripts.

To avoid confusion, the current SEAPATH class will be splitted in
SEAPATH_COMMON and SEAPATH_HOST. The VM will use the common class and
a new SEAPATH_VM class later.

This commit prepare this work by renaming SEAPATH in SEAPATH_COMMON to
be more clear.

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>