vlan on remote network
Build_debian_iso configures the main network interface of the host so that host is directly accessible via networking after deployment.
In some network topologies, this main host network interface is a trunk, and the host need to support vlan tagging to get access to the network.
This commit makes this topology supported, throught the use of the "REMOTEVLANID" configuration variable (in which the user simply has to set the vlanid number).

Signed-off-by: Florent CARLI <florent.carli@rte-france.com>