prevents FAI to upgrade to 6.2
We include the standard bookworm version of FAI in the docker image (6.0.3), but FAI upgrades the version in the NFSROOT to the one present in the project repository (fai-project.org/download).
The project recently upgraded this version to 6.2, giving an inconsistency between the build_debian_iso configuration (qualified for 6.0.3) and the actual used FAI version (6.2).

This commits prevent FAI to upgrade to 6.2 in the NFSROOT.

Signed-off-by: Florent CARLI <florent.carli@rte-france.com>