Guest Disk config: rename vg1 to vg2
It can be useful to be able to mount the guest filesystem in the host (to check on disk usage for example, or to repair something). LVM will choke if we have 2 volume groups with the same name on the same system. To solve this, since the host will still use "vg1", we rename the guest VG to "vg2".

Signed-off-by: Florent CARLI <florent.carli@rte-france.com>