build_qcow2: allows the VM disk size to be customized
The USERCUSTOMIZATION feature allow the user to configure its own partitioning for the SEAPATH VM, however, the total disk size is hardcoded in the build_qcow2.sh script.
This commit introduces a optional parameter ro specify the target total disk size for a VM (-s / --vmdisksize).
Since this script is starting to have parameters, we also add an "--help" and "--version" parameter to help a user figure out the parameters

Signed-off-by: Florent CARLI <florent.carli@rte-france.com>