SEAPATH_COMMON: add libxml2-utils
This package contains xmllint, a command line XML tool. It is used to
validate the xml file and print the error if the file is not valid.

xmllinit is used by libvirt to add more information if the xml file
is not valid. It is also used by the resource-agents VirtualDomain
during the configuration saving.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>