include cephadm binary
The cephadm packages on debian 13 is ceph v18. The advantage of using cephadm is to decorrelate the os version and the ceph version, but this is not possible if I install the first cephadm package (cephadm) in version 18, so I need to install the cephadm binary of the ceph version I will use afterwards.

Signed-off-by: Florent CARLI <florent.carli@rte-france.com>