srv_fai_config/disk_config: cases for Ceph dedicated disk installations
Rework the disk partitioning selection to be able to choose a Ceph
dedicated disk. This is done by adding a new flag "ceph_disk".
When it is enabled, the LVM Ceph partitions will not be created.

This is useful is we have a separate disk for Ceph that is not managed
by FAI. This is also useful for standalone SEAPATH installations
where Ceph is not used, in this case the "ceph_disk" flag will be
automatically added.

The "ceph_disk" flag can work with or without the "raid" flag.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>